--- nextjs/components/Channel/index.tsx
@@ -1,3 +1,0 @@
-export { Channel } from './Channel';
-export { ChannelForBots } from './ChannelForBots';
-export { NewChannelModal } from './NewChannelModal';
